// src/enums/TokenMail.ts
export enum TokenMailType {
  REGISTER_USER = "REGISTER_USER",
  REGISTER_ADMIN = "REGISTER_ADMIN",
  PASSWORD_RECOVERY_REQUEST = "PASSWORD_RECOVERY_REQUEST",
  BLOCKED_USER_ATTEMPT_LOGIN = "BLOCKED_USER_ATTEMPT_LOGIN",
  CANCEL_ACCOUNT_REQUEST = "CANCEL_ACCOUNT_REQUEST",
}

export enum TokenMailStatus {
  OPEN = "OPEN",
  EXPIRED = "EXPIRED",
  FINISHED = "FINISHED",
}